% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_add_fund_level.R
\name{fa_add_fund_level}
\alias{fa_add_fund_level}
\title{Add fund level data to attribution results}
\usage{
fa_add_fund_level(.dt, fund_data, log_convert = TRUE)
}
\arguments{
\item{.dt}{The attribution results of a portfolio.}

\item{fund_data}{The additional fund data to be added. This is the output of a \verb{fa_load_fund_*} function.}

\item{log_convert}{Should the fund returns be converted to log returns? Defaults to TRUE.}
}
\value{
A data.table of the attribution results with the additional fund level data and a new residual return component - the difference between the fund return and the implied return.
}
\description{
Add fund returns and fund yield to attribution results. A residual return component is calculated when fund returns are added measuring the difference between the bottom-up attribution returns and the top-level returns.
}
