% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_attribution.R
\name{fa_attribution}
\alias{fa_attribution}
\title{Apply the fundamental attribution calculations to a portfolio}
\usage{
fa_attribution(
  .dt,
  index_col = "index",
  id_col = "id",
  group_cols = c("metric_name", "group_id", "series_name", "series_type"),
  na_rm_start_weights = F
)
}
\arguments{
\item{.dt}{A data.table generated by a \verb{fa_load_*} function. Or, an equivalent data.table/data.frame/tibble containing the returns, yields and starting weights for each asset in the portfolio during each measurement period.}

\item{index_col}{The character name of the column containing the time index (eg period end date) for each measurement period.}

\item{id_col}{The character name of the column containing the identifier of each asset in the portfolio.}

\item{group_cols}{A character vector of columns names identifying each different portfolio and fundamental metric. This is used to group portfolios allowing multiple portfolios and fundamental metrics to be analysed together.}

\item{na_rm_start_weights}{Remove rows where start_weights are NA. Defaults to FALSE.}
}
\value{
A data.table of attribution results.
}
\description{
Calculate the fundamental attribution results for the supplied data. The data is expected to be the output from a \verb{fa_load_*} function. Portfolio returns are separated into growth, multiple, activity and intra-period components. Portfolio returns and yields implied by the underlying assets are also calculated.
}
