% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_attribution_relative.R
\name{fa_attribution_relative}
\alias{fa_attribution_relative}
\title{Calculate fundamental attribution of relative returns}
\usage{
fa_attribution_relative(strategy, benchmark)
}
\arguments{
\item{strategy}{The fundamental attribution results of the portfolio, calculated by the \code{fa_attribution} function.}

\item{benchmark}{The fundamental attribution results of the benchmark, calculated by the \code{fa_attribution} function.}
}
\value{
A data.table of the attribution results of the portfolio, the benchmark, and their relative performance.
}
\description{
Calculate the relative fundamental attribution by combining the fundamental attribution results for a portfolio and its benchmark.
}
