% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_chart_cumulative_attribution.R
\name{fa_chart_cumulative_attribution}
\alias{fa_chart_cumulative_attribution}
\title{A chart of the cumulative returns from each attribution component}
\usage{
fa_chart_cumulative_attribution(
  .dt,
  cht_caption = NULL,
  facet_scales = "free_y",
  filter_variables = c("Activity", "Growth", "Multiple", "Intra-Period", "Residual",
    "Implied Return", "Actual Return"),
  order_variables = T,
  ...
)
}
\arguments{
\item{.dt}{A data.table of attribution results, likely the output of \code{fa_chart_data_prep}.}

\item{cht_caption}{A character used in the caption of the chart. Defaults to NULL.}

\item{facet_scales}{Passed to the scales parameter of facet_wrap or facet_grid. Defaults to 'free_y'.}

\item{filter_variables}{The variables to be included in the chart, also determines the order of the variables.}

\item{order_variables}{Should the variables be ordered? If so, use the order in 'filter_variables'. Defaults to TRUE.}

\item{...}{Additional parameters that are passed to \code{geom_line} via \code{aes} within the function.}
}
\value{
A ggplot object of the cumulative return components.
}
\description{
A function to produce a simple ggplot showing the cumulative returns from each attribution component. As a ggplot object it can be edited in the usual ways.
}
