% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_load_fund_xts.R
\name{fa_load_fund_xts}
\alias{fa_load_fund_xts}
\title{Combine fund returns and fundamental-to-price ratio into data.table}
\usage{
fa_load_fund_xts(
  metric_to_price = NULL,
  fund_returns,
  metric_name = NA,
  series_name = NA,
  group_id = NA
)
}
\arguments{
\item{metric_to_price}{An xts containing the fundamental-to-price ratio of the fund. Only the first column is used.}

\item{fund_returns}{An xts containing the returns of the fund. Only the first column is used.}

\item{metric_name}{A character name of the fundamental being used.}

\item{series_name}{A character name of the strategy, this is used to match a strategy's fund level and stock level data. Defaults to the name of the first column of the fund_returns xts.}

\item{group_id}{A character name that allows multiple series to be combined, a place holder for when relative attribution is applied.}
}
\value{
A data.table of fund level data on which attribution can be run, or which can be added a strategy's stock level attribution results.
}
\description{
This function combines separate xts objects that contain fund level data into a single data.table so it can be added to attribution results calculated from the fund's underlying holdings.
}
