% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_table_log_returns_pa.R
\name{fa_table_log_returns_pa}
\alias{fa_table_log_returns_pa}
\title{A summary table of attribution return components}
\usage{
fa_table_log_returns_pa(.dt, rescale = 12, accuracy = 0.01, na_rm = T)
}
\arguments{
\item{.dt}{A data.table of attribution results.}

\item{rescale}{A number to annualise, or otherwise, rescale the returns. Defaults to 12.}

\item{accuracy}{The accuracy of the results. Defaults to 0.01.}

\item{na_rm}{Remove rows with NAs from .dt. Defaults to TRUE.}
}
\value{
A data.table summarising the returns of each attribution component.
}
\description{
Calculate the average return component for each portfolio and fundamental included in the attribution results.
}
