% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fundamental_yield_long}
\alias{fundamental_yield_long}
\title{Yields of three assets covering 13 dates}
\format{
An xts with 13 rows and 3 assets.
}
\usage{
fundamental_yield_long
}
\description{
A dataset containing the invented yields of three assets at 13 dates.
}
\keyword{datasets}
