% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fundamental_yield_short}
\alias{fundamental_yield_short}
\title{Yields of four assets covering three dates}
\format{
An xts with 3 rows and 4 assets.
}
\usage{
fundamental_yield_short
}
\description{
A dataset containing the invented yields of four assets at three dates.
}
\keyword{datasets}
