% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_long}
\alias{returns_long}
\title{Returns of three assets covering 12 periods}
\format{
An xts with 13 rows and 3 assets.
}
\usage{
returns_long
}
\description{
A dataset containing the invented returns of three assets over two periods.
}
\keyword{datasets}
