% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_short}
\alias{returns_short}
\title{Returns of four assets covering two periods}
\format{
An xts with 3 rows and 4 assets.
}
\usage{
returns_short
}
\description{
A dataset containing the invented returns of four assets over two periods.
}
\keyword{datasets}
