% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fa.R
\name{theme_fa}
\alias{theme_fa}
\title{Apply a theme to a ggplot object}
\usage{
theme_fa(base_size = 10, base_family = "", ...)
}
\arguments{
\item{base_size}{Set the text size in the ggplot.}

\item{base_family}{Set the font to use on the ggplot.}

\item{...}{Applied further theme modifications.}
}
\value{
A ggplot theme object.
}
\description{
This function applies a basic theme to a ggplot object.
}
